Fix HSPOST command syntax and add debugging information
The HSPOST command request was including an unneccesary
trailing '\r\n.\r\n' which was also being inserted by stem.
Thank you to asn for finding this issue.

This commit also increases the logging for HSPOST related control port
errors and fixes the debugging output when unknown descriptors are
received.