Add basic functional test for master descriptor publication
The functional test creates a new master key and tries to create
and upload a master descriptor using a Chutney created service as
the instance. Stem is used to fetch the descriptor from Chutney HSDir
and confirm that the master descriptor validates and contains all of
the instance's introduction points.

There is scope to expand this test to detect and assert if any warnings
or errors are output while OnionBalance is running. More tests should be
written to confirm the expected behavior when:

 * no instances are online
 * there are many (>10) instances configured
 * multiple services are configured