Add a socket which provides the service status
It can be difficult to find out the current state of the OnionBalance
daemon. This commit adds a unix domain socket which outputs the status of
the service when a client connects. By default this socket is created
at /var/run/onionbalance/control.

Thank you to Federico Ceratto for the original socket implementation.

Closes #26.