Remove dependency on "schedule" library
The schedule dependency is not available in Debian Jessie. This
prevents OnionBalance from being packaged and deployed in Debian
stable.

The scheduler process is relatively simple, so we have implemented
it directly rather than depending on an external package.

Thank you very much to Federico Ceratto for providing a new scheduler
implementation.