Correctly remove .onion TLD from instance addresses
`string.strip()` strips any of the passed characters from the string. This commit
fixes a bug where any of the characters in ".onion" would be stripped from
the end of the onion address.

The intention was to remove the string ".onion" if it was included at the end
of the provide instance address. This commit uses `.replace() to remove it.

Resolves #43.