Fix exception when an out-of-date descriptor is received
This commit fixes a bug where `None` was returned from
`Instance.update_descriptor()` rather than a boolean. If a received
instance descriptor is older than the current descriptor it was
discarded and None was returned. Instead `False` should be returned as
the boolean result it later compared in `descriptor_received()`.
s